eliminar elementos2

+++ AppNQV/views.py
@@ -113,35 +113,50 @@ def actoresFormulario(request):
 
 def BorrarPlataforma(request, pk):
    try:
-       plataforma = Plataformas.objects.get(id=pk)
-       plataforma.delete()
-       context = {'massage': 'PLATAFORMA ELIMINADA CORRECTAMENTE'}
-       return render (request, 'plataformas.html', context=context)
+       if request.method == 'GET':
+         plataforma = Plataformas.objects.get(id=pk)
+         context = {'plataforma':plataforma}
+         return render(request, 'BorrarPlataforma.html', context=context)
+       else:
+         plataforma = Plataformas.objects.get(id=pk)
+         plataforma.delete()
+         context = {'massage': 'PLATAFORMA ELIMINADA CORRECTAMENTE'}
+         return render (request, 'plataformasFormulario.html', context=context)
    except:
        context = {'error': 'Volve a intentar otra vez'}
-       return render(request, 'plataformas.html', context=context)
+       return render(request, 'plataformasFormulario.html', context=context)
 
 
 def BorrarActor(request, pk):
    try:
-       actores = Actores.objects.get(id=pk)
-       actores.delete()
-       context = {'massage3': 'ACTOR ELIMINADO/A CORRECTAMENTE'}
-       return render (request, 'actores.html', context=context)
+       if request.method == 'GET':
+         actores = Actores.objects.get(id=pk)
+         context = {'actores':actores}
+         return render(request, 'actoresBorrar.html', context=context)
+       else:
+         actores = Actores.objects.get(id=pk)
+         actores.delete()
+         context = {'massage3': 'ACTOR ELIMINADO/A CORRECTAMENTE'}
+         return render (request, 'actoresFormulario.html', context=context)
    except:
        context = {'error': 'Volve a intentar otra vez'}
-       return render(request, 'actores.html', context=context)
+       return render(request, 'actoresFormulario.html', context=context)
 
 
 def BorrarPelicula(request, pk):
    try:
-       pelicula = Peliculas.objects.get(id=pk)
-       pelicula.delete()
-       context = {'massage4': 'PELICULA ELIMINADA CORRECTAMENTE'}
-       return render (request, 'peliculas.html', context=context)
+       if request.method == 'GET':
+         pelicula = Peliculas.objects.get(id=pk)
+         context = {'pelicula':pelicula}
+         return render(request, 'peliculasBorrar.html', context=context)
+       else:
+        pelicula = Peliculas.objects.get(id=pk)
+        pelicula.delete()
+        context = {'massage4': 'PELICULA ELIMINADA CORRECTAMENTE'}
+        return render (request, 'peliculasFormulario.html', context=context)
    except:
        context = {'error': 'Volve a intentar otra vez'}
-       return render(request, 'peliculas.html', context=context)
+       return render(request, 'peliculasFormulario.html', context=context)
